actually use unsized map/array
it was a good idea until i realized they are not skippable anyway,
because the index is built linear.

if you skip a field, you also skip all of the keys in it,
and the index no longer lines up.

instead use a dumb start/end byte, which is slightly more expensive,
but ALOT simpler